# internal/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.0 DO NOT EDIT.
package gen

// Defines values for HealthStatusStatus.
const (
	Ok HealthStatusStatus = "ok"
)

// HealthStatus defines model for HealthStatus.
type HealthStatus struct {
	Status HealthStatusStatus `json:"status"`
}

// HealthStatusStatus defines model for HealthStatus.Status.
type HealthStatusStatus string

// HelloRequest defines model for HelloRequest.
type HelloRequest struct {
	Name string `json:"name"`
}

// HelloResponse defines model for HelloResponse.
type HelloResponse struct {
	Message string `json:"message"`
}

// Problem defines model for Problem.
type Problem struct {
	// Detail A human-readable explanation specific to this occurrence of the problem.
	Detail *string `json:"detail,omitempty"`

	// Errors A map of validation errors keyed by field name.
	Errors *map[string][]string `json:"errors,omitempty"`

	// Instance A URI reference that identifies the specific occurrence of the problem.
	Instance *string `json:"instance,omitempty"`

	// Status The HTTP status code generated by the origin server for this occurrence of the problem.
	Status int32 `json:"status"`

	// Title A short, human-readable summary of the problem type.
	Title string `json:"title"`

	// Type A URI reference that identifies the problem type.
	Type string `json:"type"`
}

// ProblemResponse defines model for ProblemResponse.
type ProblemResponse = Problem

// SayHelloJSONRequestBody defines body for SayHello for application/json ContentType.
type SayHelloJSONRequestBody = HelloRequest