# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway API
  version: 1.0.0
  description: |
    Schema-first gateway service providing health and greeting endpoints.
  contact:
    name: Gateway Team
    email: support@agyn.io
servers:
  - url: http://localhost:8080
tags:
  - name: system
    description: Service health and readiness endpoints.
  - name: greeting
    description: Endpoints that generate personalized greeting messages.
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      description: Returns service status and readiness indicator.
      tags:
        - system
      responses:
        '200':
          description: Service status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
              example:
                status: ok
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '422':
          $ref: '#/components/responses/ProblemResponse'
  /hello:
    post:
      operationId: sayHello
      summary: Generate a personalized greeting
      description: Responds with a greeting message that echoes the provided name.
      tags:
        - greeting
      requestBody:
        description: Greeting request payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelloRequest'
            example:
              name: Jane Doe
      responses:
        '200':
          description: Greeting response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
              example:
                message: Hello, Jane Doe
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '422':
          $ref: '#/components/responses/ProblemResponse'
components:
  schemas:
    HealthStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
    HelloRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    HelloResponse:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference that identifies the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          format: uri-reference
          description: A URI reference that identifies the specific occurrence of the problem.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: A map of validation errors keyed by field name.
      additionalProperties: false
  responses:
    ProblemResponse:
      description: RFC 7807 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'